Apply much stricter compiler options when building with GCC (#779)
* Remove --weffcpp option from setup.py.
  -Weffc++ and -Werror=effc++ are now default for GCC builds.

* Add extra warnings to GCC builds:

  * Weffc++ (error)
  * overloaded-virtual (error)
  * old-style-cast (error)
  * unused-parameter (error)

* Fixing unused warning: fix error in not passing input param
  to tree_visitor_wrapper constructor on to initialize members.

Closes #778

+++ fwdpy11/headers/fwdpy11/evolvets/evolve_generation_ts.hpp
@@ -166,7 +166,7 @@ namespace fwdpy11
                         next_index_local = offspring_node_2;
                     }
             }
-        assert(next_index_local == pop.tables->num_nodes() - 1);
+        assert(static_cast<std::size_t>(next_index_local) == pop.tables->num_nodes() - 1);
         if (next_index_local
             != static_cast<decltype(next_index_local)>(pop.tables->num_nodes()
                                                        - 1))
